+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.19.0] - 2026-01-19
+
+### Changed
+
+- Moved `zu.csv()`, `zu.stringToInteger()` and `zu.stringToUrl()` under `zu.codec`.
+They kept the same name, so `zu.csv()` became `zu.codec.csv()`.
+
 ## [0.18.0] - 2026-01-19
 
 ### Added
@@ -190,6 +197,7 @@ publication didn't succeed completely. This is a re-release.
 
 - First iteration of the library. It has JSON parsing utilities.
 
+[0.19.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.15.0...v0.16.0